v0.08 - Execution control
This time we started to work a little better with the execution of the operation.php file that will handle the operations with the database, we created a flow of access to the instructions that will be executed. We also created a validate.js file to validate the user's registration on the system. And finally, we made some small changes to the css file.